src/requirements.txt
crewai>=0.148.0
crewai-tools>=0.7.0
langchain-community>=0.0.30
openai>=1.30.1
python-dotenv>=1.0.1